refactor(generation): split dividerBlendBuilder into focused modules
Decompose the 648-LOC dividerBlendBuilder.ts into a slim orchestrator
plus three focused sibling modules. Public API preserved via
re-exports; no behavior change (28 existing tests still pass).

- dividerBlendTypes.ts (111 LOC): OuterWallCutoutInfo, DividerInfo,
  WallFaceInfo, RampZone interfaces + isPerpendicular,
  dividerTouchesWall, getWallFaceInfo + MIN_DIM/WALL_TOUCH_TOL
  constants.
- dividerBlendResolvers.ts (144 LOC): pure resolveOuterCutouts and
  collectDividers (no brepjs dependency — used by zone-computation
  helpers in the wall-pattern builder too).
- dividerBlendCuts.ts (166 LOC): buildEndTrimCut, buildRampCut,
  buildParallelTrimCut — the three brepjs cut-solid factories for the
  three blend cases.
- dividerBlendBuilder.ts (278 LOC): orchestrator buildDividerBlends +
  computeRampZones + computeDividerJunctionZones + dividerBlendFeature
  FeatureBuilder export.
